add a "sigstore --check" mode, for verifying commit signatures without git
To start with I've copy pasted lots of code from the git verify code
path, but I'll tidy that up soon. For now I'm just playing with the
ideas.

This mode is intended to be used when we want to verify the commit
signature without git being involved. The reference use case is a CI
pipeline, to calidate that the commit is from a user we trust.

Intended usage is like this:

    $ git cat-file commit HEAD | ./sigstore --check --email-domain yob.id.au --identity-provider "https://accounts.google.com"

--email-domain and --identity-provider are optional. You can leave them
off, or use one. Any commit with a valid fulcio signature:

    $ git cat-file commit HEAD | ./sigstore --check

Any commit with a valid fulcio signature where the signed identity has a
particular email domain:

    $ git cat-file commit HEAD | ./sigstore --check --email-domain yob.id.au

Any commit with a valid fulcio signature where the signed identity comes
from a particular identity provider:

    $ git cat-file commit HEAD | ./sigstore --check --identity-provider "https://accounts.google.com"
    $ git cat-file commit HEAD | ./sigstore --check --identity-provider "https://github.com/login/oauth"