check certificate validity using both fulcio certificates
There seems to be two fulcio self-signed certificates in use.

Signing is easy - we just bundle in the fulcio certificate that they
return to us.

Verifying is more complicated - we need to check both certificates to
determine if the commit was signed by a certificate that fulcio issued.

In theory we could use a single certificate store to check both certs.
However, it doesn't seem to work. I suspect possibly because the subject
of both fulcio signing certificates are identical.